+++ src/components/slide/content/multiselect-from-endpoint.jsx
@@ -23,6 +23,7 @@ function MultiselectFromEndpoint({
   optionsEndpoint,
   label = null,
   value = [],
+  disableSearch = true,
   singleSelect = false,
 }) {
   const { t } = useTranslation("common");
@@ -81,7 +82,7 @@ function MultiselectFromEndpoint({
           options={options}
           selected={getSelected(value)}
           name={name}
-          disableSearch
+          disableSearch={disableSearch}
           singleSelect={singleSelect}
           labelledBy="Select"
           overrideStrings={{
@@ -109,6 +110,7 @@ MultiselectFromEndpoint.propTypes = {
   onChange: PropTypes.func.isRequired,
   optionsEndpoint: PropTypes.string.isRequired,
   singleSelect: PropTypes.bool,
+  disableSearch: PropTypes.bool,
 };
 
 export default MultiselectFromEndpoint;