+++ CHANGELOG.md
@@ -5,6 +5,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- [#272](https://github.com/os2display/display-admin-client/pull/272)
+  - Fixed feed source selector.
 - [#264](https://github.com/os2display/display-admin-client/pull/264)
   - Added checkbox options component for use in calendar modifiers.
   - Fixed multiselect when more than one feed source of the given type is installed.